___JmolVersion="14.1.14_2014.04.15"
new feature: JANA2006 reader adds M40 molecular group support
  -- no TLS modulation
  -- no local axes (third header line of pos# record)
 



Former-commit-id: d649d21a53b618c74d053fd16daa6fdb5738ecd9

+++ Jmol/src/org/jmol/adapter/smarter/AtomIterator.java
@@ -136,8 +136,8 @@ class AtomIterator implements JmolAdapterAtomIterator {
 
 	
 	@Override
-  public int getOccupancy() {
-		return (int) (atom.foccupancy * 100);
+  public float getOccupancy() {
+		return atom.foccupancy * 100;
 	}
